script_dir/requirements.txt
# Copyright (C) 2022 Intel Corporation
# SPDX-License-Identifier: Apache-2.0
#

paramiko
python-dotenv
